additional cleanup and moving of header files
