+++ RecoBTag/SoftLepton/src/MuonTaggerMLP.cc
@@ -1,4 +1,4 @@
-#include "MuonTaggerMLP.h"
+#include "RecoBTag/SoftLepton/interface/MuonTaggerMLP.h"
 #include <cmath>
 
 double MuonTaggerMLP::Value(int index,double in0,double in1,double in2,double in3) {
