+++ RecoBTag/SoftLepton/src/MuonTaggerNoIPMLP.cc
@@ -1,4 +1,4 @@
-#include "MuonTaggerNoIPMLP.h"
+#include "RecoBTag/SoftLepton/interface/MuonTaggerNoIPMLP.h"
 #include <cmath>
 
 // no normalization
